+++ sp_alternative_cpp/include/utils.h
@@ -10,7 +10,7 @@
 
 namespace fs = std::filesystem;
 
-std::pair<std::vector<std::vector<int>>, std::unordered_map<std::string, int>>
+std::pair<std::vector<std::vector<double>>, std::unordered_map<std::string, int>>
 read_matching_matrix(const fs::path& file_path);
 
 int translate_to_matrix_index(char letter, const std::unordered_map<std::string, int>& code_to_index_dict);
